+++ lib/go-tc/deliveryservices.go
@@ -245,8 +245,8 @@ type DeliveryServiceV40 struct {
 
 	// TLSVersions is the list of explicitly supported TLS versions for cache
 	// servers serving the Delivery Service's content.
-	TLSVersions       []string               `json:"tlsVersions" db:"tls_versions"`
-	GeoLimitCountries *GeoLimitCountriesType `json:"geoLimitCountries"`
+	TLSVersions       []string              `json:"tlsVersions" db:"tls_versions"`
+	GeoLimitCountries GeoLimitCountriesType `json:"geoLimitCountries"`
 }
 
 // DeliveryServiceV4 is a Delivery Service as it appears in version 4 of the
@@ -578,8 +578,10 @@ type DeliveryServiceNullableV11 struct {
 	DeliveryServiceRemovedFieldsV11
 }
 
+// GeoLimitCountriesType is the type alias that is used to represent the GeoLimitCountries attribute of the DeliveryService struct
 type GeoLimitCountriesType []string
 
+// UnmarshalJSON will unmarshal a byte slice into type GeoLimitCountriesType
 func (g *GeoLimitCountriesType) UnmarshalJSON(data []byte) error {
 	var err error
 	var initial = make([]string, 0)
@@ -589,9 +591,10 @@ func (g *GeoLimitCountriesType) UnmarshalJSON(data []byte) error {
 			return err
 		}
 		if strings.Contains(initialStr, ",") {
-			return errors.New("array of strings need to be passed in as an array, and not a string")
+			initial = strings.Split(initialStr, ",")
+		} else {
+			initial = append(initial, initialStr)
 		}
-		initial = append(initial, initialStr)
 	}
 
 	if initial == nil || len(initial) == 0 {
@@ -603,6 +606,7 @@ func (g *GeoLimitCountriesType) UnmarshalJSON(data []byte) error {
 
 }
 
+// MarshalJSON will marshal a GeoLimitCountriesType into a byte slice
 func (g GeoLimitCountriesType) MarshalJSON() ([]byte, error) {
 	arr := ([]string)(g)
 	return json.Marshal(arr)
@@ -845,11 +849,9 @@ func (ds *DeliveryServiceV4) RemoveLD1AndLD2() DeliveryServiceV4 {
 // DowngradeToV3 converts the 4.x DS to a 3.x DS.
 func (ds *DeliveryServiceV4) DowngradeToV3() DeliveryServiceNullableV30 {
 	nullableFields := ds.DeliveryServiceNullableFieldsV11
-	if ds.GeoLimitCountries != nil {
-		geoLimitCountries := ([]string)(*ds.GeoLimitCountries)
-		geo := strings.Join(geoLimitCountries, ",")
-		nullableFields.GeoLimitCountries = &geo
-	}
+	geoLimitCountries := ([]string)(ds.GeoLimitCountries)
+	geo := strings.Join(geoLimitCountries, ",")
+	nullableFields.GeoLimitCountries = &geo
 	return DeliveryServiceNullableV30{
 		DeliveryServiceV30: DeliveryServiceV30{
 			DeliveryServiceNullableV15: DeliveryServiceNullableV15{
@@ -888,7 +890,7 @@ func (ds *DeliveryServiceNullableV30) UpgradeToV4() DeliveryServiceV4 {
 		DeliveryServiceFieldsV13:         ds.DeliveryServiceFieldsV13,
 		DeliveryServiceNullableFieldsV11: ds.DeliveryServiceNullableFieldsV11,
 		TLSVersions:                      nil,
-		GeoLimitCountries:                &geo,
+		GeoLimitCountries:                geo,
 	}
 }